requirements.txt
numpy==1.24.3
Pillow==10.0.1
Flask==2.3.3
redis==4.6.0
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
